Move probe extraction to the parent class
