+++ pftools/snc_conv.py
@@ -34,7 +34,6 @@ class sncConversion(PFConversion):
         self.mesh = None
         self.data = None
         self.vtk_object = None
-        self.probe = None
         
         if self.verbose:
             print('PF file format is: {0:s}'.format(self.format))
@@ -601,42 +600,3 @@ class sncConversion(PFConversion):
         
         self.probe[probe_name] = DataFrame(data=data)
         
-
-    def export_temporal_data(self,casename,dirout,delimiter=' ',index=False,
-                             extension='txt'):
-        """Function to export probe temporal data to a text file. 
-        All quantities will be written in SI units
-
-        Parameters
-        ----------
-        casename : string
-            Name assiciated to the present probe conversion. 
-            The casename is used to build the output file name
-            temporal_<casename>.<extension>
-        dirout : string
-            Absolute path/relative path where the converted file will be written
-        delimiter : char
-            Field delimiter (default is space).
-            If comma ',' is specified the file extension will be 'csv'
-        index : bool
-            Append index of rows as the first column in the text file
-        extension : string
-            Extension of the text file (by default txt)
-
-        """
-
-        import os.path
-
-        if delimiter == ',':
-            ext = 'csv'
-        else:
-            ext = extension
-            
-        if self.probe is None:
-            raise RuntimeError('No probe to export, please use extract_probe')
-        
-        for probe_name in self.probe.keys():
-            outFile = os.path.join(dirout,'temporal_{0:s}_{2:s}.{1:s}'.format(
-                        casename,ext,probe_name))
-            print("Exporting in ascii column format:\n  ->  {0:s}".format(outFile))
-            self.probe[probe_name].to_csv(outFile,sep=delimiter,index=index)
